Error: 카테고리 돔 렌더링은 되는데 이벤트 등록이 안 됨
