+++ index.html
@@ -74,55 +74,7 @@
         </menu>
       </div>
     </article>
-    <main id="main"></main>
-    <!-- <header id="header">
-      <div class="category">
-        <div class="category__button" id="category-button">
-          <i class="fas fa-bars"></i>
-          <span>카테고리</span>
-        </div>
-        <div class="none category__layer" id="category-layer">
-          <ul class="category__list">
-            <li><span>패션의류/잡화</span></li>
-            <li><span>뷰티</span></li>
-            <li><span>식품</span></li>
-            <li><span>출산/유아동</span></li>
-            <li><span>생활용품</span></li>
-            <li><span>주방용품</span></li>
-            <li><span>홈인테리어</span></li>
-            <li><span>도서/음반/DVD</span></li>
-            <li><span>스포츠/레저</span></li>
-            <li><span>문구/오피스</span></li>
-            <li><span>반려동물용품</span></li>
-            <li><span>헬스/건강식품</span></li>
-            <li><span>여행/티켓</span></li>
-          </ul>
-          <div class="category__depth">
-            <ul class="category__list-inner">
-              <li><span>유아/어린이</span></li>
-              <li><span>소설/에세이/시</span></li>
-              <li><span>초중고참고서</span></li>
-              <li><span>가정 살림</span></li>
-              <li><span>건강 취미</span></li>
-              <li><span>경제 경영</span></li>
-              <li><span>과학/공학</span></li>
-              <li><span>국어/외국어/사전</span></li>
-              <li><span>대학교재</span></li>
-              <li><span>만화/라이트노벨</span></li>
-            </ul>
-          </div>
-          <div class="category__depth">
-            <ul class="category__list-inner">
-              <li><span>유아/어린이</span></li>
-              <li><span>소설/에세이/시</span></li>
-              <li><span>초중고참고서</span></li>
-              <li><span>가정 살림</span></li>
-              <li><span>건강 취미</span></li>
-            </ul>
-          </div>
-        </div>
-      </div>
-    </header> -->
+    <div id="root"></div>
     <script type="module" src="src/index.js"></script>
   </body>
 </html>
