+++ src/layout/Header.js
@@ -1,8 +1,8 @@
+import Category from '../components/Category.js';
 import HtmlElement from '../utils/HtmlElement.js';
-import { findTargetIdElement } from '../utils/manuplateDOM.js';
 
-function Header(htmlTag, $parent) {
-  HtmlElement.call(this, htmlTag, $parent);
+function Header(htmlTag) {
+  HtmlElement.call(this, htmlTag);
   this.setTemplate();
   this.render();
   this.setEvent();
@@ -13,74 +13,9 @@ Header.prototype.constructor = Header;
 // Object.setPrototypeOf(Header.prototype, HtmlElement.prototype);
 
 Header.prototype.setTemplate = function () {
-  // 나중에 데이터도 넣어보기
-  this.template = template;
+  this.$element.id = 'header';
+  const category = new Category('div');
+  this.template = category.$element.innerHTML;
 };
-
-Header.prototype.setEvent = function () {
-  const $categoryButton = findTargetIdElement(this.$element, 'category-button');
-  $categoryButton.addEventListener('click', handleCategoryButton.bind(this));
-};
-
-function handleCategoryButton(event) {
-  const $categoryLayer = findTargetIdElement(this.$element, 'category-layer');
-  const layerClassList = $categoryLayer.classList;
-  if (layerClassList.contains('none')) {
-    $categoryLayer.classList.remove('none');
-    $categoryLayer.classList.add('show');
-  } else {
-    $categoryLayer.classList.remove('show');
-    $categoryLayer.classList.add('none');
-  }
-}
-
-const template = `<header id="header">
-<div class="category">
-  <div class="category__button" id="category-button">
-    <i class="fas fa-bars"></i>
-    <span>카테고리</span>
-  </div>
-  <div class="none category__layer" id="category-layer">
-    <ul class="category__list">
-      <li><span>패션의류/잡화</span></li>
-      <li><span>뷰티</span></li>
-      <li><span>식품</span></li>
-      <li><span>출산/유아동</span></li>
-      <li><span>생활용품</span></li>
-      <li><span>주방용품</span></li>
-      <li><span>홈인테리어</span></li>
-      <li><span>도서/음반/DVD</span></li>
-      <li><span>스포츠/레저</span></li>
-      <li><span>문구/오피스</span></li>
-      <li><span>반려동물용품</span></li>
-      <li><span>헬스/건강식품</span></li>
-      <li><span>여행/티켓</span></li>
-    </ul>
-    <div class="category__depth">
-      <ul class="category__list-inner">
-        <li><span>유아/어린이</span></li>
-        <li><span>소설/에세이/시</span></li>
-        <li><span>초중고참고서</span></li>
-        <li><span>가정 살림</span></li>
-        <li><span>건강 취미</span></li>
-        <li><span>경제 경영</span></li>
-        <li><span>과학/공학</span></li>
-        <li><span>국어/외국어/사전</span></li>
-        <li><span>대학교재</span></li>
-        <li><span>만화/라이트노벨</span></li>
-      </ul>
-    </div>
-    <div class="category__depth">
-      <ul class="category__list-inner">
-        <li><span>유아/어린이</span></li>
-        <li><span>소설/에세이/시</span></li>
-        <li><span>초중고참고서</span></li>
-        <li><span>가정 살림</span></li>
-        <li><span>건강 취미</span></li>
-      </ul>
-    </div>
-  </div>
-</div>
-</header>`;
 
 export default Header;
