+++ src/utils/HtmlElement.js
@@ -1,18 +1,14 @@
-import { findTargetIdElement } from './manuplateDOM.js';
-
 function HtmlElement(htmlTag) {
-  this.$parent = findTargetIdElement(document, 'main');
   this.$element = document.createElement(htmlTag);
   this.template = '';
 }
 
 HtmlElement.prototype.setTemplate = function () {
-  return ``;
+  // child요소.render한 데이터를 this.$element에 appendChild
 };
 
 HtmlElement.prototype.render = function () {
   this.$element.innerHTML = this.template;
-  this.$parent.appendChild(this.$element);
 };
 
 HtmlElement.prototype.setEvent = function () {};
