+++ style/scss/components/_category.scss
@@ -23,6 +23,7 @@
   height: 400px;
   border: 2px solid $main-grey-light;
   box-shadow: -1px 4px 5px rgba($main-black, 0.3);
+  background-color: $main-white;
   z-index: 1;
   display: flex;
 }
