+++ style/scss/style.scss
@@ -8,5 +8,5 @@
   display: none;
 }
 .show {
-  display: show;
+  display: flex;
 }
